pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "haip-poc"
version = "0.1.0"
description = "AI agent that interviews game testers via voice overlay"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-speech>=2.36.1",
    "google-cloud-texttospeech>=2.34.0",
    "google-genai>=1.62.0",
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pyqt6>=6.10.2",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]

[project.scripts]
haip = "haip_poc.main:main"

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]

[tool.uv]
package = true

[tool.hatchling.build.targets.wheel]
packages = ["haip_poc"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"
